requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
Flask-WTF==1.2.1
python-dotenv==1.0.0
email-validator==2.1.0
cryptography==46.0.4